requirements.txt
requests==2.22.0
Django==3.1.3
Flask==2.0.2
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
dataclasses==0.6
sqlalchemy=1.4.31
paho-mqtt=5.0
azure-cognitiveservices-vision-computervision=-0.9.0
pillow=9.1.0